// frontend/src/lib/backend/apiV1/schema.d.ts
export interface components {
    schemas: {
        RsDataVoid: {
            code: string;
            msg: string;
            data: Record<string, never>;
        };
        CommentModifyReqBody: {
            content?: string;
        };
        Empty: Record<string, never>;
        RsDataEmpty: {
            code: string;
            msg: string;
            data: components["schemas"]["Empty"];
        };
        PostModifyReqBody: {
            title: string;
            content: string;
            published?: boolean;
            listed?: boolean;
        };
        PostWithContentDto: {
            /** Format: int64 */
            id: number;
            /** Format: date-time */
            createdDate: string;
            /** Format: date-time */
            modifiedDate: string;
            title: string;
            /** Format: int64 */
            authorId: number;
            content: string;
            authorName: string;
            published: boolean;
            listed: boolean;
        };
        RsDataPostWithContentDto: {
            code: string;
            msg: string;
            data: components["schemas"]["PostWithContentDto"];
        };
        WriteReqBody: {
            title: string;
            content: string;
            published?: boolean;
            listed?: boolean;
        };
        LoginReqBody: {
            username: string;
            password: string;
        };
        LoginResBody: {
            item: components["schemas"]["MemberDto"];
            apiKey: string;
            accessToken: string;
        };
        MemberDto: {
            /** Format: int64 */
            id: number;
            /** Format: date-time */
            createdDate: string;
            /** Format: date-time */
            modifiedDate: string;
            nickname: string;
        };
        RsDataLoginResBody: {
            code: string;
            msg: string;
            data: components["schemas"]["LoginResBody"];
        };
        JoinReqBody: {
            username: string;
            password: string;
            nickname: string;
        };
        RsDataMemberDto: {
            code: string;
            msg: string;
            data: components["schemas"]["MemberDto"];
        };
        PageDto: {
            items: components["schemas"]["PostDto"][];
            /** Format: int32 */
            totalPages: number;
            /** Format: int32 */
            totalItems: number;
            /** Format: int32 */
            currentPageNo: number;
            /** Format: int32 */
            pageSize: number;
        };
        PostDto: {
            /** Format: int64 */
            id: number;
            /** Format: date-time */
            createdDate: string;
            /** Format: date-time */
            modifiedDate: string;
            title: string;
            /** Format: int64 */
            authorId: number;
            authorName: string;
            published: boolean;
            listed: boolean;
        };
        RsDataPageDto: {
            code: string;
            msg: string;
            data: components["schemas"]["PageDto"];
        };
        CommentDto: {
            /** Format: int64 */
            id: number;
            content: string;
            /** Format: int64 */
            postId: number;
            /** Format: int64 */
            authorId: number;
            authorName: string;
            /** Format: date-time */
            createdTime: string;
            /** Format: date-time */
            modifiedTime: string;
        };
        RsDataStatisticsResBody: {
            code: string;
            msg: string;
            data: components["schemas"]["StatisticsResBody"];
        };
        StatisticsResBody: {
            /** Format: int64 */
            postCount: number;
            /** Format: int64 */
            postPublishedCount: number;
            /** Format: int64 */
            postListedCount: number;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}